chore(sanity): move versions menu to versions folder
